# section2_api/openapi.json
{"openapi":"3.1.0","info":{"title":"Powerball API","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/export/first-position":{"get":{"summary":"Export First Position","operationId":"export_first_position_api_export_first_position_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":69,"minimum":1,"default":69,"title":"Limit"}},{"name":"source","in":"query","required":false,"schema":{"type":"string","default":"auto","title":"Source"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/export/first-position.xlsx":{"get":{"summary":"Export First Position Excel","operationId":"export_first_position_excel_api_export_first_position_xlsx_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":69,"minimum":1,"default":69,"title":"Limit"}},{"name":"source","in":"query","required":false,"schema":{"type":"string","default":"auto","title":"Source"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history/by-position":{"get":{"summary":"History By Position","operationId":"history_by_position_api_history_by_position_get","parameters":[{"name":"pos","in":"query","required":true,"schema":{"type":"integer","title":"Pos"}},{"name":"num","in":"query","required":true,"schema":{"type":"integer","title":"Num"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":5000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history/filter":{"get":{"summary":"History Filter","operationId":"history_filter_api_history_filter_get","parameters":[{"name":"white1","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White1"}},{"name":"white2","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White2"}},{"name":"white3","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White3"}},{"name":"white4","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White4"}},{"name":"white5","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White5"}},{"name":"powerball","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Powerball"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date To"}},{"name":"complete","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Complete"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"draw_date","title":"Sort"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"asc","title":"Direction"}},{"name":"output","in":"query","required":false,"schema":{"type":"string","default":"json","title":"Output"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":5000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history/filter/or":{"get":{"summary":"History Filter Or","operationId":"history_filter_or_api_history_filter_or_get","parameters":[{"name":"white1","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White1"}},{"name":"white2","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White2"}},{"name":"white3","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White3"}},{"name":"white4","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White4"}},{"name":"white5","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White5"}},{"name":"powerball","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Powerball"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date To"}},{"name":"complete","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Complete"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"draw_date","title":"Sort"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"asc","title":"Direction"}},{"name":"output","in":"query","required":false,"schema":{"type":"string","default":"lines","title":"Output"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":5000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history/filter/atleast":{"get":{"summary":"History Filter Atleast","operationId":"history_filter_atleast_api_history_filter_atleast_get","parameters":[{"name":"white1","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White1"}},{"name":"white2","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White2"}},{"name":"white3","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White3"}},{"name":"white4","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White4"}},{"name":"white5","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White5"}},{"name":"powerball","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Powerball"}},{"name":"min_match","in":"query","required":false,"schema":{"type":"integer","default":2,"title":"Min Match"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date To"}},{"name":"complete","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Complete"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"score","title":"Sort"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc","title":"Direction"}},{"name":"output","in":"query","required":false,"schema":{"type":"string","default":"lines","title":"Output"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":5000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/future/quickpick":{"post":{"summary":"Api Future Quickpick","operationId":"api_future_quickpick_api_future_quickpick_post","parameters":[{"name":"n","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":1,"title":"N"}},{"name":"draw_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Draw Date"}},{"name":"seed","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seed"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/future/filter":{"get":{"summary":"Api Future Filter","operationId":"api_future_filter_api_future_filter_get","parameters":[{"name":"white1","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White1"}},{"name":"white2","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White2"}},{"name":"white3","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White3"}},{"name":"white4","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White4"}},{"name":"white5","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White5"}},{"name":"powerball","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Powerball"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date To"}},{"name":"complete","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Complete"}},{"name":"output","in":"query","required":false,"schema":{"type":"string","default":"json","title":"Output"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"created_at","title":"Sort"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc","title":"Direction"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/future/filter/or":{"get":{"summary":"Api Future Filter Or","operationId":"api_future_filter_or_api_future_filter_or_get","parameters":[{"name":"white1","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White1"}},{"name":"white2","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White2"}},{"name":"white3","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White3"}},{"name":"white4","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White4"}},{"name":"white5","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White5"}},{"name":"powerball","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Powerball"}},{"name":"complete","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Complete"}},{"name":"output","in":"query","required":false,"schema":{"type":"string","default":"json","title":"Output"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"created_at","title":"Sort"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc","title":"Direction"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/future/filter/atleast":{"get":{"summary":"Api Future Filter Atleast","operationId":"api_future_filter_atleast_api_future_filter_atleast_get","parameters":[{"name":"white1","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White1"}},{"name":"white2","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White2"}},{"name":"white3","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White3"}},{"name":"white4","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White4"}},{"name":"white5","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White5"}},{"name":"powerball","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Powerball"}},{"name":"min_match","in":"query","required":false,"schema":{"type":"integer","default":2,"title":"Min Match"}},{"name":"complete","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Complete"}},{"name":"output","in":"query","required":false,"schema":{"type":"string","default":"json","title":"Output"}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"score","title":"Sort"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"desc","title":"Direction"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/future/import-excel":{"post":{"summary":"Api Future Import Excel","operationId":"api_future_import_excel_api_future_import_excel_post","parameters":[{"name":"sheet_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sheet Name"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_api_future_import_excel_api_future_import_excel_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/future/quickpick/unique":{"post":{"summary":"Api Future Quickpick Unique","operationId":"api_future_quickpick_unique_api_future_quickpick_unique_post","parameters":[{"name":"n","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":1,"title":"N"}},{"name":"draw_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Draw Date"}},{"name":"seed","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seed"}},{"name":"white1","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White1"}},{"name":"white2","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White2"}},{"name":"white3","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White3"}},{"name":"white4","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White4"}},{"name":"white5","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"White5"}},{"name":"powerball","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Powerball"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_api_future_import_excel_api_future_import_excel_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_api_future_import_excel_api_future_import_excel_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}